libkefir.h: fix description of index for rule insertion
If a filter already has a rule at the index provided when inserting a
new rule, the new one does NOT overwrite the old one. Instead, it is
inserted just before, thus shifting the index of the old rule and all
the following ones in the list.

Let's fix the description.

Also mention the possibility to have negative indices, to indicate the
count should go backwards from the end of the list.

Reported by @pcasconnetronome.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>