libkefir, compile: make struct bpf_object args "const" when possible
Following a recent update from libbpf, many functions from this library
now use "const" to indicate that the struct bpf_object for which they
take a pointer as an argument is not modified during the execution.
Therefore we can make our own function taking pointers to "const" struct
bpf_object (as long as they remain unmodified in the function) without
triggering compiler warnings. Let's do it, to make explicit to users,
when relevant, that those objects are not modified by libkefir
functions.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>