proggen: use ethertype to check IPv4/6 values were collected
When we try to see if a rule matches against a packet, we compare the
values in the key with the values in the rule. But if the values in the
key were not found in the packet (e.g. IPv4 source address not found in
IPv6 packet), and if for example the rule provided by the user has a
pattern at 0, then a false positive match can be returned.

To work around this issue we need to make sure that the values compared
were actually collected before comparing them with the fields in the
rule.

This fix will be done over several commit. The present one makes sure
that IPv4 and IPv6 values were collected when comparing items to
patterns based on IPv4 or IPv6 elements: the ethertype is evaluated each
time before doing the comparison, to check that the context is relevant.
Also make ethertype part of the "key", rather than passing it around to
all the functions that need it (key is passed to the same functions
already).

Reported by @kuba-moo.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>