proggen: flag L4 processing
Flag L4 processing to make sure the values in the key have been
collected (even if at zero) before comparing them to what is in the
rule.

We need to make a specific case for L4 "data" (first four bytes of L4
header), for which we have no guaranty that this is a TCP/UDP/SCTP
header and which may be shorter than any of those headers.

When matching on either IPv4/IPv6, remove the check on ethertype being
either of those IP protocols, as having processed L4 implies that we
have an IP packet.

Reported by @kuba-moo.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>